+++ packages/genesys/packages/react-native-components/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Versions
 
+# 3.4.4
+
+[QrScanner] fix: fix `StatusBar` moving the app up and down when opening the qr scanner on android.
+
+# 3.3.11
+
+[Backdrop] fix: fix `avoidKeyboard` on Android.
+
+# 3.3.10
+
+[package.json] chore: update react-native-components
+
 # 3.3.7
 
 [Backdrop] fix: Backdrop not filling entire screen (see)[https://github.com/react-native-modal/react-native-modal#the-backdrop-is-not-completely-filledcovered-on-some-android-devices-galaxy-for-one]